Remove TEXT wire parser replacing it with YAML
Split out common YAML formatting code
closes #411
