+++ src/test/java/net/openhft/chronicle/wire/issue/JSON222Test.java
@@ -119,8 +119,6 @@ public class JSON222Test extends WireTestCommon {
                 if (expected.contains("\r\n"))
                     expected = expected.replaceAll("\r\n", "\n");
                 String actual = bytes2.toString();
-                // TODO FIX Reduce redundant padding around [ ] in JSON #392
-                actual = actual.replaceAll("\\[ ", "[");
                 assertEquals(expected, actual);
             }
             // if (fail)
