+++ src/test/resources/nst_files/e-15.yaml
@@ -8,6 +8,6 @@
     "5": [
       "6"
     ],
-    ? [7]: ""
+    ? [ 7]: ""
   }
 ]
