+++ src/test/resources/nst_files/e-173.yaml
