+++ src/test/resources/nst_files/e-176.yaml
