+++ src/test/resources/nst_files/e-24.yaml
