+++ src/test/resources/nst_files/e-26.yaml
